PR feedback

+++ src/Ludo.Common/Models/GameOrchestrator.cs
@@ -36,15 +36,15 @@ public class GameOrchestrator
     throw new NotImplementedException();
   }
 
-  public int[] DetermineStartingPlayer(int[] rolls)
+  public byte[] DetermineStartingPlayer(int[] rolls)
   {
     int highestRoll = rolls.Max();
 
-    List<int> highestRollers = [];
+    List<byte> highestRollers = [];
 
     for (int i = 0; i < rolls.Length; i++)
       if (rolls[i] == highestRoll)
-        highestRollers.Add(i);
+        highestRollers.Add((byte) i);
 
     return highestRollers.ToArray();
   }
